requirements.txt
yt_dlp
python-telegram-bot